phonebook_api/requirements.txt
requests
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt
django-phonenumber-field[phonenumberslite]
django-filter